Ensure a failing tests actually breaks the build.

+++ ros_pkg_with_tests/scripts/test/test_double.py
@@ -9,7 +9,7 @@ from scripts.double import double
 class TestCase(unittest.TestCase):
 
     def test_four(self):
-        ret = double(4)
+        ret = double(2)
         self.assertEqual(8, ret)
 
     def test_two(self):
